Patches by Alain: use <signal.h> instead of <sys/signal.h>, truncate pid_t before using as ping_ident.

+++ libicmp/icmp_echo.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 1998 Free Software Foundation, Inc.
+/* Copyright (C) 1998, 2001 Free Software Foundation, Inc.
 
    This file is part of GNU Inetutils.
 
@@ -23,7 +23,7 @@
 
 #include <sys/socket.h>
 #include <sys/time.h>
-#include <sys/signal.h>
+#include <signal.h>
 
 #include <netinet/in_systm.h>
 #include <netinet/in.h>
@@ -36,7 +36,7 @@ int
 icmp_generic_encode(u_char *buffer, size_t bufsize, int type, int ident, int seqno)
 {
   icmphdr_t *icmp;
-  
+
   if (bufsize < 8)
     return -1;
   icmp = (icmphdr_t *)buffer;
@@ -54,24 +54,24 @@ int
 icmp_generic_decode(u_char *buffer, size_t bufsize,
 		    struct ip **ipp, icmphdr_t **icmpp)
 {
-  int hlen;
+  size_t hlen;
   u_short cksum;
   struct ip *ip;
   icmphdr_t *icmp;
-  
+
   /* IP header */
   ip = (struct ip*) buffer;
   hlen = ip->ip_hl << 2;
   if (bufsize < hlen + ICMP_MINLEN)
     return -1;
-  
+
   /* ICMP header */
   icmp = (icmphdr_t*)(buffer + hlen);
 
   /* Prepare return values */
   *ipp = ip;
   *icmpp = icmp;
-  
+
   /* Recompute checksum */
   cksum = icmp->icmp_cksum;
   icmp->icmp_cksum = 0;
